backend/projex-api/requirements.txt
# FastAPI + ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth + Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
pyotp==2.9.0
qrcode[pil]==8.0

# Validation
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Redis + Caching
redis[hiredis]==5.2.1

# Task queue
celery[redis]==5.4.0

# HTTP client
httpx==0.28.1

# Meilisearch
meilisearch==0.31.6

# MinIO / S3
minio==7.2.12

# Monitoring
prometheus-client==0.21.1

# Dev/Test
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1

# Linting
ruff==0.8.6
